+++ src/main/java/ru/mihaly4/vkmusictransfer/client/IVkClient.java
@@ -1,8 +1,8 @@
 package ru.mihaly4.vkmusictransfer.client;
 
 public interface IVkClient {
-    String fromProfile(int id, int offset);
-    String fromCommunity(String id, int offset);
-    String getRemixSid();
+    String fromAudio(int id, int offset);
+    String fromWall(String id, int offset);
+    Boolean login(String username, String password);
     int getUid();
 }