+++ src/main/java/ru/mihaly4/vkmusictransfer/config/PackageConfig.java
@@ -5,8 +5,6 @@ import java.util.List;
 public class PackageConfig {
     private String tgbToken = "";
     private String tgbUsername = "";
-    private String vkRemixSid = "";
-    private int vkUid = 0;
     private List<String> trustedTgbUsers;
 
     public String getTgbToken() {
@@ -23,22 +21,6 @@ public class PackageConfig {
 
     public void setTgbUsername(String tgbUsername) {
         this.tgbUsername = tgbUsername;
-    }
-
-    public String getVkRemixSid() {
-        return vkRemixSid;
-    }
-
-    public void setVkRemixSid(String vkRemixSid) {
-        this.vkRemixSid = vkRemixSid;
-    }
-
-    public int getVkUid() {
-        return vkUid;
-    }
-
-    public void setVkUid(int vkUid) {
-        this.vkUid = vkUid;
     }
 
     public List<String> getTrustedTgbUsers() {