+++ src/main/java/ru/mihaly4/vkmusictransfer/di/BotComponent.java
@@ -3,8 +3,10 @@ package ru.mihaly4.vkmusictransfer.di;
 import dagger.Component;
 import ru.mihaly4.vkmusictransfer.Bot;
 
+import javax.inject.Singleton;
+
 @Component(modules = BotModule.class)
-@SharedScope
+@Singleton
 public interface BotComponent {
     Bot makeBot();
 }