+++ src/main/java/ru/mihaly4/vkmusictransfer/parser/LoginParser.java
@@ -1,0 +1,33 @@
+package ru.mihaly4.vkmusictransfer.parser;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+/**
+ * @since 1.1.0
+ */
+public class LoginParser {
+    public static String parseUrl(String data) {
+        Pattern p = Pattern.compile("action=\"([^\"]+)");
+
+        Matcher matches = p.matcher(data);
+
+        return matches.find() ? matches.group(1) : "";
+    }
+
+    public static int parseUid(String data) {
+        Pattern p = Pattern.compile("class=\"ip_user_link[\\s\\S]+href=\"/id([\\d]+)");
+
+        Matcher matches = p.matcher(data);
+
+        return matches.find() ? Integer.valueOf(matches.group(1)) : 0;
+    }
+
+    public static String parseRemixSid(String data) {
+        Pattern p = Pattern.compile("remixsid=([^;]+)");
+
+        Matcher matches = p.matcher(data);
+
+        return matches.find() ? matches.group(1) : "";
+    }
+}