+++ src/test/java/ru/mihaly4/vkmusictransfer/repository/VkRepositoryTest.java
@@ -23,7 +23,7 @@ public class VkRepositoryTest {
 
         final CountDownLatch signal = new CountDownLatch(1);
 
-        repository.findAllByProfile(123456789).thenAccept(action -> {
+        repository.findAllByAudio(123456789).thenAccept(action -> {
             links.putAll(action);
 
             signal.countDown();
@@ -42,7 +42,7 @@ public class VkRepositoryTest {
 
         final CountDownLatch signal = new CountDownLatch(1);
 
-        repository.findAllByCommunity("example").thenAccept(action -> {
+        repository.findAllByWall("example").thenAccept(action -> {
             links.putAll(action);
 
             signal.countDown();
@@ -55,12 +55,12 @@ public class VkRepositoryTest {
 
     private static class VkClient implements IVkClient {
         @Override
-        public String fromProfile(int id, int offset) {
+        public String fromAudio(int id, int offset) {
             if (offset != 0) {
                 return "";
             }
 
-            URL fixture = getClass().getResource("/fixture/profile.html");
+            URL fixture = getClass().getResource("/fixture/audio.html");
 
             try {
                 return Resources.toString(fixture, Charsets.UTF_8);
@@ -70,12 +70,12 @@ public class VkRepositoryTest {
         }
 
         @Override
-        public String fromCommunity(String id, int offset) {
+        public String fromWall(String id, int offset) {
             if (offset != 0) {
                 return "";
             }
 
-            URL fixture = getClass().getResource("/fixture/community.html");
+            URL fixture = getClass().getResource("/fixture/wall.html");
 
             try {
                 return Resources.toString(fixture, Charsets.UTF_8);
@@ -85,13 +85,13 @@ public class VkRepositoryTest {
         }
 
         @Override
-        public String getRemixSid() {
-            return "";
+        public Boolean login(String username, String password) {
+            return true;
         }
 
         @Override
         public int getUid() {
-            return 444529088; // need for to decode
+            return 444529088;
         }
     }
 }